feat: Implement automatic snack name extraction from Coupang URLs
We've changed the SnackCommand to automatically extract the snack name from given Coupang URLs. This is achieved by introducing a parseQuery method in a new class, CoupangUrlParser, which is now used in place of directly reading a provided 'snack' string. As a result of these updates, the 'snack' option has been removed from snack.json.

This change will remove the need for users to manually input the snack name, making the snack request process more streamlined and user-friendly. We've also accounted for potential invalid URLs by handling URISyntaxException, which prompts a user-friendly error message asking for user's recheck.

Tests for the new parseQuery method are added in the CoupangUrlParserTest class to ensure the correctness of the new feature.